Fix navigation issues & imports
