--- app/index.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-import { signedIn } from "./router";
-
-export default class App extends React.Component {
-	render() {
-		console.log('hakeem is here');
-		return (
-			signedIn
-		)
-	}
-}
