+++ app/layout/SignIn.tsx
@@ -2,8 +2,13 @@ import React from 'react';
 import { View } from 'react-native';
 import { Card, Button, FormLabel, FormInput } from 'react-native-elements';
 import { onSignIn } from '../auth';
+interface Props {
+	navigation?: any; // Question mark indicates prop is optional
+}
 
-export class SignIn extends React.Component {
+interface State {
+}
+export class SignIn extends React.Component<Props, State> {
 	render() {
 		return (
 			<View style={{ paddingVertical: 20 }}>
