+++ index.ios.ts
@@ -1,3 +1,1 @@
-import { AppRegistry } from 'react-native';
-import App from './App';
-AppRegistry.registerComponent('gossip', () => App);
+import './app/index';
